Cleanup and comment certs package
