+++ internal/log/perms.go
@@ -6,10 +6,6 @@ package log
 /*
    team - Embedded teamserver for Go programs and CLI applications
    Copyright (C) 2023 Reeflective
-   This program is free software: you can redistribute it and/or modify
-   it under the terms of the GNU General Public License as published by
-   the Free Software Foundation, either version 3 of the License, or
-   (at your option) any later version.
 
    This program is free software: you can redistribute it and/or modify
    it under the terms of the GNU General Public License as published by
@@ -24,6 +20,7 @@ package log
    You should have received a copy of the GNU General Public License
    along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
+
 import (
 	"fmt"
 	"os"
